Move commands into dedicated module
