+++ src/commands/delete.rs
