+++ src/commands/info.rs
