+++ src/lib.rs
@@ -4,28 +4,18 @@ extern crate quickcheck;
 
 pub mod capabilities;
 pub mod cgroups;
-pub mod command;
+pub mod commands;
 pub mod container;
-pub mod create;
 pub mod dbus;
-pub mod delete;
-pub mod exec;
-pub mod info;
-pub mod kill;
-pub mod list;
 pub mod logger;
 pub mod namespaces;
 pub mod notify_socket;
-pub mod pause;
 pub mod pipe;
 pub mod process;
-pub mod resume;
 pub mod rootfs;
 pub mod rootless;
 pub mod signal;
-pub mod spec_json;
-pub mod start;
-pub mod state;
 pub mod stdio;
+pub mod syscall;
 pub mod tty;
 pub mod utils;
